use multi-stage build to minimize size of Docker image
The changes in this commit reduce the Docker image size from >1 GB to 205 MB. Docker multi-stage builds were used to achieve this. In the first stage, the core ta-lib library is compiled, and the Python ta-lib package is installed. Then, in the second stage, the compiled core ta-lib library and the Python wheel files are copied over and installed. The second stage also uses the `python:VERSION-slim` Docker image, which is much smaller than the `python:VERSION` image, because it does not include any compilers or other software not necessary at runtime.

The build and run procedure is unchanged. Use `docker build --tag ta-lib .` to build and `docker run --rm -it ta-lib bash` to run.

Another change introduced by the commit is the option of skipping testing during image building. By default, testing is not skipped, but it can be skipped with `docker uild --tag ts-lib --build-arg RUN_TESTS=0 .`.